% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boolean_to_french.R
\name{boolean_to_french}
\alias{boolean_to_french}
\title{Return the word VRAI if TRUE and FAUX if FALSE}
\usage{
boolean_to_french(boolean)
}
\arguments{
\item{boolean}{A boolean TRUE or FALSE}
}
\description{
Return the word VRAI if TRUE and FAUX if FALSE
}
