% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/century_rank.R
\name{century_rank}
\alias{century_rank}
\title{Finds the day of the century for a particular year}
\usage{
century_rank(year, as_a_number = TRUE)
}
\arguments{
\item{year}{A year as a number}

\item{as_a_number}{If TRUE, the day as a number if FALSE, the day as a string}
}
\value{
The day for that century
}
\description{
Finds the day of the century for a particular year
}
