% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doomsday.R
\name{doomsday}
\alias{doomsday}
\title{Find the doomsday (as a number) for a particular year}
\usage{
doomsday(year, as_a_number = TRUE)
}
\arguments{
\item{year}{A year as a number}

\item{as_a_number}{If TRUE, the day as a number if FALSE, the day as a string}
}
\value{
The doomsday of that year as a number
}
\description{
Find the doomsday (as a number) for a particular year
}
\examples{
doomsday(2021)
doomsday(1928)
doomsday(1978, as_a_number = FALSE)
}
