% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doomsdays_df.R
\name{doomsdays_df}
\alias{doomsdays_df}
\title{Makes a tibble where the columns consists of the days of the week
and the lines give the year}
\usage{
doomsdays_df(start, end)
}
\arguments{
\item{start}{A year to start the dataframe}

\item{end}{A year to end the dataframe}
}
\value{
A tibble containing the years associated with their doomsdays
}
\description{
Makes a tibble where the columns consists of the days of the week
and the lines give the year
}
