% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eleven_on_odd.R
\name{eleven_on_odd}
\alias{eleven_on_odd}
\title{Use the Eleven On Odd method to find doomsday for a specific year}
\usage{
eleven_on_odd(infrasecular)
}
\arguments{
\item{infrasecular}{A two digit number}
}
\value{
The gap between the year anchor day and doomsday
}
\description{
Use the Eleven On Odd method to find doomsday for a specific year
}
