% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_weekday.R
\name{find_weekday}
\alias{find_weekday}
\title{Finds the day of the week for a particular date}
\usage{
find_weekday(my_date, as_a_number = TRUE)
}
\arguments{
\item{my_date}{A date in the format defined in lubridate, ymd(20211221)}

\item{as_a_number}{A boolean to choose to give the day as a number or as a word, default = TRUE}
}
\value{
A number or a word that gives the day of the week
}
\description{
Finds the day of the week for a particular date
}
