% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_leap_year.R
\name{is_leap_year}
\alias{is_leap_year}
\title{Finds if a year is a leap year}
\usage{
is_leap_year(year)
}
\arguments{
\item{year}{The year we want to test for leap year}
}
\value{
A boolean, TRUE is a leap year, FALSE if not
}
\description{
Finds if a year is a leap year
}
\examples{
is_leap_year(1900)
is_leap_year(2000)
}
