% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_leap_year_mad.R
\name{is_leap_year_mad}
\alias{is_leap_year_mad}
\title{Finds if a year is a leap year (function used for testing leap year algorithm)}
\usage{
is_leap_year_mad(year)
}
\arguments{
\item{year}{The year we want to test for leap year}
}
\value{
A boolean, TRUE is a leap year, FALSE if not
}
\description{
Finds if a year is a leap year (function used for testing leap year algorithm)
}
