% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_to_day.R
\name{num_to_day}
\alias{num_to_day}
\title{Turn a number to a day of the week}
\usage{
num_to_day(num)
}
\arguments{
\item{num}{A number between 0 and 6}
}
\value{
The day of the week
}
\description{
Turn a number to a day of the week
}
\examples{
num_to_day(1)
num_to_day(5)
}
