% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_to_month.R
\name{num_to_month}
\alias{num_to_month}
\title{Turn a number to a month}
\usage{
num_to_month(num)
}
\arguments{
\item{num}{A number between 1 and 12}
}
\value{
The name of the month
}
\description{
Turn a number to a month
}
