% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_date.R
\name{pretty_date}
\alias{pretty_date}
\title{A function to pretty print a date}
\usage{
pretty_date(my_date)
}
\arguments{
\item{my_date}{A date in the format of YEAR-MONTH-DAY as a string}
}
\value{
A string that prints the date in a pretty manner
}
\description{
A function to pretty print a date
}
\examples{
pretty_date(lubridate::ymd(20220113))
pretty_date(lubridate::ymd(19780702))
}
