% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_eleven_on_odd.R
\name{print_eleven_on_odd}
\alias{print_eleven_on_odd}
\title{Print the Eleven On Odd method to find doomsday for a specific year}
\usage{
print_eleven_on_odd(my_infrasecular, itemize = TRUE)
}
\arguments{
\item{infrasecular}{A two digit number}
}
\value{
A string...
}
\description{
Print the Eleven On Odd method to find doomsday for a specific year
}
